fix: thread flicker
threads would flicker on receiving new notes when auto scroll is
enabled. This is because auto scroll is doing the same actions a
user would be doing, which creates children ThreadNodes. These
new children have an empty virtual list when they are created for
the first time. Since it's a new list, there isn't a last width
set, so the flicker gets triggered if hide_on_resize is enabled.

The easiest solution is to simply disable hide_on_resize. The
flicker isn't very noticable to me, but if this isn't acceptable
we can propagate the width to the child nodes (and corresponding
lists) so we could still use the hide_on_resize but not hide on
every thread open.

This flicker was actually happening every time we opened a thread
or reply for the first time, but it wasn't noticible because we
would always be navigating while it happens

Signed-off-by: kernelkind <kernelkind@gmail.com>